pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "dynvision"
version = "0.0.1"
description = "Dynamical Vision Modeling"
authors = [{ name = "Inkognito" }]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "ruamel.yaml ~=0.18.3",
    "python-dotenv ~=1.0.1",
    "scipy ~=1.9",                                       # formerly 1.7.3
    "snakemake ~=8.10.0",
    "snakemake-executor-plugin-cluster-generic ~=1.0.0",
    "pandas ~=1.2.0",
    "scikit-learn ~=1.1.0",
    "scikit-image ~=0.20.0",
    "torch >=2.2.0",
    "torchvision >=0.16.0",
    "torchsummary ~=1.5.0",
    "opencv-python ~=4.9.0",
    "matplotlib ~=3.5.1",
    "seaborn ~=0.13.0",
    "jupyterlab ~=4.0.0",
    "ipympl ~=0.7.0",
    "tqdm ~=4.66.0",
    "wandb",
    "pytorch-lightning",
    "tensorboard",
    "cupy ~= 13.3.0",
    "numba ~= 0.60.0",
    "cudatoolkit ~= 11.3",
    "libjpeg-turbo ~= 3.0.0",
    "opencv ~= 4.10.0",
    "pkg-config ~= 0.29.2",
    "ffcv ~= 1.0.2",
]


[tool]
python_requires = "~= 3.11"

[project.optional-dependencies]
dev = [
    "black",
    "bump2version",
    "flake8",
    "mypy",
    "pre-commit",
    "pydocstyle",
    "pylint",
    "pytest",
    "pytest-black",
    "pytest-cov",
    "pytest-flake8",
    "pytest-mypy",
    "pytest-pydocstyle",
    "pytest-pylint",
    "pytest-xdist",
]
doc = [
    "sphinx>=3.3.0",
    "numpydoc>=1.1.0",
    "sphinx-tabs>=1.3.0",
    "sphinx-argparse>=0.4.0",
    "sphinx-autodoc-typehints",
    "pygments",
    "recommonmark",
]


[tool.black]
line-length = 87
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.flake8]
ignore = "E203"
